Scalafmt: change configuration.
currently, scalafmt is configured to introduce newlines when there are
at least 2 parameters in a method declaration, or call.
This can be excessive, so we raise that number to 3.

Also, we make sure that the `implicit` keyword in a list of parameters
appears before, on the same line, as the first implicit parameter.